+++ linuxdoom-1.10/p_doors.c
@@ -411,7 +411,7 @@ EV_VerticalDoor
 	
     // if the sector has an active thinker, use it
     sec = sides[ line->sidenum[side^1]] .sector;
-    secnum = sec-sectors;
+    secnum = (int)(sec-sectors);
 
     if (sec->specialdata)
     {
